improved ncr precision

+++ snippets/math/combinatorics.cpp
@@ -6,8 +6,8 @@ ull nPr(ull n, ull r) {
 }
 
 ull nCr(ull n, ull r) {
-	double v = 1;
-	for (int i = 1; i <= r; i++)
+	long double v = 1;
+	for (ull i = 1; i <= r; i++)
 		v = v * (n-r+i) /i;
 	return (ull)(v + 0.001);
 }
